+++ packages/terra-application-links/examples/test-examples/config/testShortConfig.js
@@ -1,0 +1,8 @@
+const testShortConfig = [
+  { path: '/test1', text: 'test 1' },
+  { path: '/test2', text: 'test 2' },
+  { path: '/test3', text: 'test 3' },
+  { path: '/test4', text: 'test 4' },
+];
+
+export default testShortConfig;