+++ packages/terra-application-name/examples/test-examples/ApplicationMenuNameDefault.example.jsx
@@ -3,9 +3,11 @@ import Image from 'terra-image';
 import ApplicationMenuName from '../../lib/ApplicationMenuName';
 
 export default () => (
-  <ApplicationMenuName
-    id="default"
-    title="Title"
-    accessory={<Image alt="terra accessory" variant="rounded" src="https://github.com/cerner/terra-core/raw/master/terra.png" height="26px" width="26px" />}
-  />
+  <div style={{ backgroundColor: 'green', width: '100%' }}>
+    <ApplicationMenuName
+      id="default"
+      title="Title"
+      accessory={<Image alt="terra accessory" variant="rounded" src="https://github.com/cerner/terra-core/raw/master/terra.png" height="26px" width="26px" />}
+    />
+  </div>
 );