+++ packages/terra-layout/CHANGELOG.md
@@ -4,9 +4,13 @@ Changelog
 Unreleased
 ----------
 
+### Changed
+* Returned promises from toggleMenu/togglePin
+* Updated menu overlay styles.
+
 2.1.0 - (March 6, 2018)
 ------------------
-###Changed
+### Changed
 * Remove the hover functionality from the layout panel at the medium breakpoint.
 
 ### Added